op-e2e: Rearrange fault proof tests
Moves fault proof tests to their own package and splits them into separate files based on the game type.
