+++ op-e2e/faultproofs/alphabet_test.go
@@ -1,0 +1,177 @@
+package faultproofs
+
+import (
+	"context"
+	"testing"
+	"time"
+
+	op_e2e "github.com/ethereum-optimism/optimism/op-e2e"
+	"github.com/ethereum-optimism/optimism/op-e2e/e2eutils/challenger"
+	"github.com/ethereum-optimism/optimism/op-e2e/e2eutils/disputegame"
+	"github.com/ethereum-optimism/optimism/op-e2e/e2eutils/wait"
+	"github.com/stretchr/testify/require"
+)
+
+func TestChallengerCompleteDisputeGame(t *testing.T) {
+	op_e2e.InitParallel(t, op_e2e.UseExecutor(1))
+
+	tests := []struct {
+		name              string
+		rootClaimAlphabet string
+		otherAlphabet     string
+		expectedResult    disputegame.Status
+		expectStep        bool
+	}{
+		{
+			name:              "ChallengerWins_DefenseStep",
+			rootClaimAlphabet: "abcdexyz",
+			otherAlphabet:     disputegame.CorrectAlphabet,
+			expectedResult:    disputegame.StatusChallengerWins,
+			expectStep:        true,
+		},
+		{
+			name:              "DefenderWins_DefenseStep",
+			rootClaimAlphabet: disputegame.CorrectAlphabet,
+			otherAlphabet:     "abcdexyz",
+			expectedResult:    disputegame.StatusDefenderWins,
+			expectStep:        false,
+		},
+		{
+			name:              "ChallengerWins_AttackStep",
+			rootClaimAlphabet: "abcdefghzyx",
+			otherAlphabet:     disputegame.CorrectAlphabet,
+			expectedResult:    disputegame.StatusChallengerWins,
+			expectStep:        true,
+		},
+		{
+			name:              "DefenderWins_AttackStep",
+			rootClaimAlphabet: disputegame.CorrectAlphabet,
+			otherAlphabet:     "abcdexyz",
+			expectedResult:    disputegame.StatusDefenderWins,
+			expectStep:        false,
+		},
+		{
+			name:              "DefenderIncorrectAtTraceZero",
+			rootClaimAlphabet: "zyxwvut",
+			otherAlphabet:     disputegame.CorrectAlphabet,
+			expectedResult:    disputegame.StatusChallengerWins,
+			expectStep:        true,
+		},
+		{
+			name:              "ChallengerIncorrectAtTraceZero",
+			rootClaimAlphabet: disputegame.CorrectAlphabet,
+			otherAlphabet:     "zyxwvut",
+			expectedResult:    disputegame.StatusDefenderWins,
+			expectStep:        false,
+		},
+		{
+			name:              "DefenderIncorrectAtLastTraceIndex",
+			rootClaimAlphabet: "abcdefghijklmnoz",
+			otherAlphabet:     disputegame.CorrectAlphabet,
+			expectedResult:    disputegame.StatusChallengerWins,
+			expectStep:        true,
+		},
+		{
+			name:              "ChallengerIncorrectAtLastTraceIndex",
+			rootClaimAlphabet: disputegame.CorrectAlphabet,
+			otherAlphabet:     "abcdefghijklmnoz",
+			expectedResult:    disputegame.StatusDefenderWins,
+			expectStep:        false,
+		},
+	}
+	for _, test := range tests {
+		test := test
+		t.Run(test.name, func(t *testing.T) {
+			op_e2e.InitParallel(t, op_e2e.UseExecutor(1))
+
+			ctx := context.Background()
+			sys, l1Client := startFaultDisputeSystem(t)
+			t.Cleanup(sys.Close)
+
+			disputeGameFactory := disputegame.NewFactoryHelper(t, ctx, sys.Cfg.L1Deployments, l1Client)
+			game := disputeGameFactory.StartAlphabetGame(ctx, test.rootClaimAlphabet)
+			require.NotNil(t, game)
+			gameDuration := game.GameDuration(ctx)
+
+			game.StartChallenger(ctx, sys.NodeEndpoint("l1"), "Defender",
+				challenger.WithAgreeProposedOutput(false),
+				challenger.WithPrivKey(sys.Cfg.Secrets.Mallory),
+			)
+
+			game.StartChallenger(ctx, sys.NodeEndpoint("l1"), "Challenger",
+				// Agree with the proposed output, so disagree with the root claim
+				challenger.WithAgreeProposedOutput(true),
+				challenger.WithAlphabet(test.otherAlphabet),
+				challenger.WithPrivKey(sys.Cfg.Secrets.Alice),
+			)
+
+			// Wait for a claim at the maximum depth that has been countered to indicate we're ready to resolve the game
+			game.WaitForClaimAtMaxDepth(ctx, test.expectStep)
+
+			sys.TimeTravelClock.AdvanceTime(gameDuration)
+			require.NoError(t, wait.ForNextBlock(ctx, l1Client))
+
+			game.WaitForInactivity(ctx, 10, true)
+			game.LogGameData(ctx)
+			require.EqualValues(t, test.expectedResult, game.Status(ctx))
+		})
+	}
+}
+
+func TestChallengerCompleteExhaustiveDisputeGame(t *testing.T) {
+	op_e2e.InitParallel(t, op_e2e.UseExecutor(1))
+
+	testCase := func(t *testing.T, isRootCorrect bool) {
+		ctx := context.Background()
+		sys, l1Client := startFaultDisputeSystem(t)
+		t.Cleanup(sys.Close)
+
+		disputeGameFactory := disputegame.NewFactoryHelper(t, ctx, sys.Cfg.L1Deployments, l1Client)
+		rootClaimedAlphabet := disputegame.CorrectAlphabet
+		if !isRootCorrect {
+			rootClaimedAlphabet = "abcdexyz"
+		}
+		game := disputeGameFactory.StartAlphabetGame(ctx, rootClaimedAlphabet)
+		require.NotNil(t, game)
+		gameDuration := game.GameDuration(ctx)
+
+		// Start honest challenger
+		game.StartChallenger(ctx, sys.NodeEndpoint("l1"), "Challenger",
+			challenger.WithAgreeProposedOutput(!isRootCorrect),
+			challenger.WithAlphabet(disputegame.CorrectAlphabet),
+			challenger.WithPrivKey(sys.Cfg.Secrets.Alice),
+			// Ensures the challenger responds to all claims before test timeout
+			challenger.WithPollInterval(time.Millisecond*400),
+		)
+
+		// Start dishonest challenger
+		dishonestHelper := game.CreateDishonestHelper(disputegame.CorrectAlphabet, 4, !isRootCorrect)
+		dishonestHelper.ExhaustDishonestClaims(ctx)
+
+		// Wait until we've reached max depth before checking for inactivity
+		game.WaitForClaimAtDepth(ctx, int(game.MaxDepth(ctx)))
+
+		// Wait for 4 blocks of no challenger responses. The challenger may still be stepping on invalid claims at max depth
+		game.WaitForInactivity(ctx, 4, false)
+
+		sys.TimeTravelClock.AdvanceTime(gameDuration)
+		require.NoError(t, wait.ForNextBlock(ctx, l1Client))
+
+		expectedStatus := disputegame.StatusChallengerWins
+		if isRootCorrect {
+			expectedStatus = disputegame.StatusDefenderWins
+		}
+		game.WaitForInactivity(ctx, 10, true)
+		game.LogGameData(ctx)
+		require.EqualValues(t, expectedStatus, game.Status(ctx))
+	}
+
+	t.Run("RootCorrect", func(t *testing.T) {
+		op_e2e.InitParallel(t, op_e2e.UseExecutor(1))
+		testCase(t, true)
+	})
+	t.Run("RootIncorrect", func(t *testing.T) {
+		op_e2e.InitParallel(t, op_e2e.UseExecutor(1))
+		testCase(t, false)
+	})
+}
